# docs/swagger.yaml
definitions:
...
  ErrorResponse:
    properties:
      error:
        type: string
      errors:
        items:
          $ref: '#/definitions/view.ApiError'
        type: array
    type: object
...
  MessageResponse:
    properties:
      message:
        type: string
    type: object
...
  view.ApiError:
    description: validation error details
    properties:
      enums:
        description: available options incase of field's payload is enums
        items:
          type: string
        type: array
      field:
        description: the field cause the error
        type: string
      msg:
        description: error message
        type: string
    type: object
...
info:
  contact:
    email: benjamin@d.foundation
    name: Nam Nguyen
    url: https://d.foundation
  description: This is api document for fortress project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: FORTRESS API DOCUMENT
  version: v0.1.39
paths:
...
  /notion/audiences:
    get:
      consumes:
      - application/json
      description: Get list audiences from DF Audience
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list audiences from DF Audience
      tags:
      - Notion
...
  /notion/digests:
    get:
      consumes:
      - application/json
      description: Get list digests from DF Internal Digest
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list digests from DF Internal Digest
      tags:
      - Notion
...
  /notion/hiring-positions:
    get:
      consumes:
      - application/json
      description: Get list hiring from DF Dwarves Hiring
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list hiring from DF Dwarves Hiring
      tags:
      - Notion
  /notion/issues:
    get:
      consumes:
      - application/json
      description: Get list issues from DF Issues & Resolution Log
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list issues from DF Issues & Resolution Log
      tags:
      - Notion
  /notion/memos:
    get:
      consumes:
      - application/json
      description: Get list memos from DF Memos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list memos from DF Memos
      tags:
      - Notion
  /notion/projects/milestones:
    get:
      consumes:
      - application/json
      description: Get list  project milestones
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list  project milestones
      tags:
      - Notion
...
  /notion/tech-radars:
    get:
      consumes:
      - application/json
      description: Get list items from DF TechRadar
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ErrorResponse'
      summary: Get list items from DF TechRadar
      tags:
      - Notion
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"